+++ src/MicroEngine.Game/Scenes/Demos/GraphicsDemo.cs
@@ -14,7 +14,7 @@ namespace MicroEngine.Game.Scenes.Demos;
 public sealed class GraphicsDemo : Scene
 {
     private readonly IInputBackend _inputBackend;
-    private readonly IRenderBackend _renderBackend;
+    private readonly IRenderBackend2D _renderBackend;
     private readonly ILogger _logger;
     private readonly ResourceCache<ITexture> _textureCache;
     private readonly Random _random;